Move NAT into a separate subfolder and namespace
Move NAT.cs into a separate folder and into a separate namespace Pax.Examples.Nat in preparation for refactoring
