+++ examples/Examples.csproj
@@ -32,7 +32,7 @@
     <Compile Include="Hub.cs" />
     <Compile Include="LearningSwitch.cs" />
     <Compile Include="Mirror.cs" />
-    <Compile Include="NAT.cs" />
+    <Compile Include="Nat\NAT.cs" />
     <Compile Include="Test.cs" />
   </ItemGroup>
 <!--
